[ci] Split up integration.shard dart_data_asset_test.dart
